Cleanup after rename.
